オブジェクト管理を分離
